% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaXdesign.R
\name{plaXpower}
\alias{plaXpower}
\alias{print.plaXpower}
\title{Sample Size and Power Comparisons of Placebo Crossover and Standard Two-Arm Designs with Two Periods of Follow-up}
\usage{
plaXpower(
  theta1,
  theta2,
  ve1,
  ve2,
  method = c("analytical", "simulation"),
  alpha = 0.025,
  iter = 1000,
  seed = NULL
)

\method{print}{plaXpower}(x, digits = 2, ...)
}
\arguments{
\item{theta1}{a numeric value specifying the Poisson mean event count in the original placebo arm in period 1 in the standard design. This mean count is assumed in all scenarios described by \code{ve1} and \code{ve2}.}

\item{theta2}{a numeric value specifying the Poisson mean event count in the original placebo arm in period 2 in the standard design. This mean count is assumed in all scenarios described by \code{ve1} and \code{ve2}.}

\item{ve1}{a numeric vector specifying different scenarios of vaccine efficacy in period 1 defined as 1 minus the ratio (original vaccine/original placebo) of mean event counts}

\item{ve2}{a numeric vector specifying different scenarios of vaccine efficacy in period 2 defined as 1 minus the ratio (original vaccine/original placebo) of mean event counts. Vectors \code{ve1} and \code{ve2} must be of equal length, with matching components of \code{ve1} and \code{ve2} treated as a single trial scenario.}

\item{method}{a character string specifying the computational method. The default option is \code{"analytical"}, which assumes that the mean event counts and VE levels in both arms and time periods and in both designs are known. The alternative option is \code{"simulation"}, which first randomly samples event counts from Poisson distributions defined by \code{theta1}, \code{theta2}, \code{ve1}, and \code{ve2}, and then uses the data to estimate VE and the variance. Only the first character of the method is necessary.}

\item{alpha}{one-sided nominal significance level (0.025 by default)}

\item{iter}{the number of Monte-Carlo iterations applied for the \code{"simulation"} method (\eqn{10^3} by default)}

\item{seed}{a seed of the random number generator supplied to \code{set.seed} for reproducibility}

\item{x}{an object of class \code{plaXpower} from a call of \code{plaXpower}}

\item{digits}{the number of decimal places to use for printing (2 by default)}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{plaXpower}, which is a data frame with the following columns:
\itemize{
\item \code{ve1}: repeats the input argument \code{ve1}
\item \code{ve2}: repeats the input argument \code{ve2}
\item \code{ssRatioWaneXstd}: the sample size ratio (placebo crossover/standard) that ensures the same power to detect waning VE
\item \code{ssRatioHarmXstd}: the sample size ratio (placebo crossover/standard) that ensures the same power to detect vaccine harm
\item \code{powerWaneX}: power to detect waning VE in the placebo crossover design
\item \code{powerWaneStd}: power to detect waning VE in the standard non-crossover design
\item \code{powerHarmX}: power to detect vaccine harm in the placebo crossover design
\item \code{powerHarmStd}: power to detect vaccine harm in the standard non-crossover design
}
}
\description{
Considering two periods of participant follow-up with associated levels of vaccine efficacy \eqn{VE_1} and \eqn{VE_2}, \code{plaXpower} computes for both a placebo crossover and standard non-crossover design (i) power to reject the null hypothesis \eqn{H_0: VE_1 = VE_2} in favor of the one-sided alternative hypothesis of waning VE, i.e., \eqn{H_1: VE_1 > VE_2}, (ii) power to reject the null hypothesis \eqn{H_0: VE_2 \ge 0} in favor of the one-sided alternative hypothesis of vaccine harm, i.e., \eqn{H_1: VE_2 < 0}, and (iii) for each hypothesis test, the sample size ratio (placebo crossover/standard) required to achieve the same power. The event of interest is assumed to be rare, with the event count following a Poisson distribution in each arm and time period. The vaccine administered in period 2 is assumed to have equal VE as in period 1.
}
\details{
The Wald test statistic used for evaluating waning VE in the placebo crossover design is
\deqn{\frac{\{\log(1 - \widehat{VE}_2) - \log(1 - \widehat{VE}_1) \}}{\sqrt{\frac{1}{\widehat{\theta}_2 (1-\widehat{VE}_1)} + \frac{1}{\widehat{\theta}_2 (1-\widehat{VE}_2)}}}}.
}
\examples{
# analytical method
tab <- plaXpower(theta1=100,
                 theta2=100,
                 ve1=rep(0.8, 5),
                 ve2=c(0.8, 0.5, 0.2, -0.1, -0.4))

# pretty printing of 'tab' with 2 decimal places (default)
tab

# pretty printing of 'tab' with 3 decimal places
print(tab, digits=3)

# unrounded real numbers in 'tab' for, e.g., plotting of a power curve
tab$powerWaneX

# simulation method
tab <- plaXpower(theta1=100,
                 theta2=100,
                 ve1=rep(0.8, 5),
                 ve2=c(0.8, 0.5, 0.2, -0.1, -0.4),
                 method="s",
                 iter=100,
                 seed=19283)

}
